Implement spells and behaviours
Both the Mage and Archer will try to get out of the reach of enemies
if they're next to them and the leader isn't engaging that monster.
HealingSpell has also been created, and now the first 3 spells are of
this type. The Mage has spells equipped as their secondary item and
will now attempt to heal any hero who's health has dipped below half.
To accompany this, I've added another sprite sheet so we can have
sprites to indicate healing, as well as the sound effect and +HP text